feat: move get tables call so it does it on datasource change
